requirements.txt
python-telegram-bot[job-queue]==22.3
aiohttp==3.9.5
websockets==12.0
python-dotenv==1.0.1
